requirements.txt
flask
flask-cors
sqlalchemy
pandas
python-multipart
pyjwt
bcrypt
openpyxl
gunicorn
